distress signal admin button (#3186)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

A simple button that creates a announcement on whatever ship your
hovering over!

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/a38409fb-9530-47ba-9c77-cb28e8e75645)

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/6230d9da-246c-41ee-8e08-7453a275c400)

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/4317a4a6-0a6b-4550-89df-47a7248d31ca)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Should save admins a lot of time

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: New admin button to launch a distress announcement quicker.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->