Shetland wiring fix (#3237)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fixes a minor wiring issue on the Shetland, where the input and output
power lines are shorted together.

Before:

![shetlandShort_circled](https://github.com/user-attachments/assets/6769c666-7a6f-4b9e-b4cf-866229a992cd)

![roundstartShetlandPowerMonitor_NoFix](https://github.com/user-attachments/assets/55ef1250-f05c-47bf-b198-97b8c803bf17)

After:

![shetlandShort_fixed](https://github.com/user-attachments/assets/f9e9280c-92db-4451-b3cd-e8556c107aad)

![roundstartShetlandPowerMonitor_Fixed](https://github.com/user-attachments/assets/a1ee706a-8fa3-4984-a5d7-5fa7549d855a)

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

This short bypasses the SMES units aboard and more or less renders them
useless. This change should fix this, and allow them to work as they're
supposed to.

## Changelog

:cl:
fix: fixed a short on the Shetland
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->