Giving-the-PGF-some-more-love (PGF Cold Protections and Crying Sun Changes) (#3210)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR includes the following changes:

- Slightly expands and adds to the bridge to be more comfortable to
operate from.
- Adds a security camera console to the bridge to allow for full bridge
monitoring
- Sets the Crying Sun and the the Nail's turrets to be set to lethal by
default
- The radiation crate in stowage now contains radiation suits for the
Navy Engineers.
- An additional standard vest added to marine officer's locker, should
the coated variant not be desired.
- Equipment in the marine armory has been adjusted to reflect its
purpose of equipping the marines and the marines only.
- Silkenweave jackets and Betiz hats now offer zero armor, but offer
cold protection, for people moving onto frosty planets
- Adds a dresser to the crew quarters
- Removes two Ion engines and replaces them with combustion engines for
a total of four combustion engines and two ion engines. Reconfigures the
engine rooms to reflect this change.

Picture of the changed bridge:
![Screenshot 2024-08-02
181827](https://github.com/user-attachments/assets/47b234f1-eb8a-422d-9853-66fbdf383927)

Picture of the entire ship with areas:
![Screenshot 2024-08-02
181846](https://github.com/user-attachments/assets/d3a6dd90-9766-4978-8a00-21d8d00f9258)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The changes to equipment on board the ship are good for players as it
provides more options. Players will now have trade offs to consider when
chosing their equipment in adverse conditions. Because of the PGF's
strict uniform protocols, providing these options to players also allows
them to enjoy different styles and looks whilst remaining lore friendly.

The changes to the ship's layout on the bridge is in response to the
innability of the ship's command staff to be everywhere at once. By
including more space to work on the bridge and a security camera
console, the captain now has access to essential equipment in close
proximity to their primary work station, the helm.

Finally, with the addition of combustion engines, including ships that
utilize them more heavily makes for interesting variaton and challenge.
Increasing engineering interaction.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a standard vest to the armory office locker
add: Added a Security camera console to the bridge
balance: Silkenweave jackets and Betzu hats now provide cold protection
but no armor
balance: Crying Sun now has 4 combustion/2 ion engines instead of 2/4
fix: Fixes turrets on the Crying Sun and its subshuttle, the Nail.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>