Box Hospital Ship engineering power fix and dress up (#3261)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This fixes the box's issues with power connections crossing in a corner,
causing the SMES to feedback on itself. It also dressed it up just a bit
so its slightly less jank.

Old Layout:


![image](https://github.com/user-attachments/assets/544dacd9-9671-4810-a61a-54ac66d714b6)


New Layout:


![image](https://github.com/user-attachments/assets/245c3624-9ae3-4f05-8470-96061f0b55ca)


## Why It's Good For The Game

Fixes are great.

## Changelog

:cl:
fix: Adjusted the Box Hospital Ship's engineering layout so that it's
less jank, and the power system doesn't loop on itself anymore.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->