Minor fix to 10mm rubber rounds (#3325)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fixes a typo in the typepath for 10mm rubber rounds.

## Why It's Good For The Game

This should allow for 10mm rubber rounds to be obtainable from their
respective ammo box now.

## Changelog

:cl:
fix: fixed typepath for 10mm rubber rounds
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->