changes human generic names to be improper (e.g. the human man) + changes mothperson to moth (#3321)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![imagen](https://github.com/user-attachments/assets/eeef1198-1f3b-4de8-8ba1-f1f67d97adb6)
title

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
* mothperson person was getting a bit jarring
*
![imagen](https://github.com/user-attachments/assets/ec18a121-e8ab-4d7f-a7f6-5cf8108828ac)
< this wont happen



<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
spellcheck: Human generic names (the ones that show up when you examine
someone) will be prefixed consistently
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->