Adds some of the new Frontiersmen weapons to the Black Market (#3332)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Adds the Mauler, Spitter and Pounder and their respective magazines to
the Black Market

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

They're illicit, pirate weapons, and the Spitter itself was made in mind
for black market usage in it's descriptions, so they're fitting
additions to the black market, and gives these interesting weapons a
place to be bought and seen.

## Changelog

:cl:
add: Adds the Mauler, Spitter, and Pounder to the black market
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->