Nerfs hollowpoint rounds to bring them under limb crit threshold (#3334)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Nerfs most hollowpoint rounds by 5 damage. Nerfs 8mm hollowpoint by 6
damage for a total of 49 damage.

## Why It's Good For The Game

This is mostly because .357 Hollowpoint does exactly 50 damage, which
crits an unarmored limb (which is most players' limbs) in a single shot.
The nerf doesn't hit HP so hard it isn't worth it, because you're still
doing the same damage as weapons much higher up on the totem pole, but
it ensures the only weapon that can crit a limb in a single shot is a
.45-70 firing hollowpoint or explosive ammo. And sniper rifles like the
Scout, if those ever get something like hollowpoint.

## Changelog

:cl:
balance: Reduced hollowpoint damage by 5 (and 8mm Hollowpoint by 6)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->